requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
click>=8.1.0
pandas>=1.5.0
streamlit>=1.15.0
pydeck>=0.8.0 